+++ theme-canary/pages/privacy.php
@@ -1,0 +1,5 @@
+<?php
+defined('MYAAC') or die('Direct access not allowed!');
+
+require PLUGINS . 'theme-canary/themes/canary/pages/privacy.php';
+